fix: Nanvix platform fixes for Python 3.12
- Fix PY_SSIZE_T_MAX signedness: Nanvix Newlib defines SSIZE_MAX as
  (__SIZE_MAX__ >> 1) retaining unsigned type, causing signed/unsigned
  comparison bugs in slice operations (e.g., list[::2] returned wrong
  results). Use signed literal 0x7FFFFFFFL on Nanvix.

- Fix %zd format specifier: Newlib printf does not support %zd. Replace
  with %ld/(long) casts in bytesobject.c and use %ld-based format tables
  in unicodeobject.c under __nanvix__ guard.

- Enable dynamic loading: Add CCSHARED=-fPIC and LDSHARED=$(CC) -shared
  for Nanvix in configure.ac. Remove -static from CFLAGS (keep in LDFLAGS).

- Enable additional modules: Re-enable _bz2, pyexpat, _elementtree, and
  _zoneinfo (previously disabled for Nanvix). Add libbz2 to LIBS.

- Add INSTALL_PREFIX support: Allow CI to override --prefix to /sysroot
  so release artifacts don't contain ephemeral runner paths.

+++ Objects/bytesobject.c
@@ -273,7 +273,7 @@ PyBytes_FromFormatV(const char *format, va_list vargs)
                 sprintf(buffer, "%ld", va_arg(vargs, long));
             }
             else if (size_tflag) {
-                sprintf(buffer, "%zd", va_arg(vargs, Py_ssize_t));
+                sprintf(buffer, "%ld", (long)va_arg(vargs, Py_ssize_t));
             }
             else {
                 sprintf(buffer, "%d", va_arg(vargs, int));
@@ -287,7 +287,7 @@ PyBytes_FromFormatV(const char *format, va_list vargs)
                 sprintf(buffer, "%lu", va_arg(vargs, unsigned long));
             }
             else if (size_tflag) {
-                sprintf(buffer, "%zu", va_arg(vargs, size_t));
+                sprintf(buffer, "%lu", (unsigned long)va_arg(vargs, size_t));
             }
             else {
                 sprintf(buffer, "%u", va_arg(vargs, unsigned int));
